+++ csmall-product/src/main/java/com/tqq/csmall/product/service/impl/BrandServiceImpl.java
@@ -3,7 +3,6 @@ package com.tqq.csmall.product.service.impl;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.tqq.csmall.product.ex.ServiceException;
 import com.tqq.csmall.product.mapper.BrandMapper;
-import com.tqq.csmall.product.pojo.entity.AttributeTemplate;
 import com.tqq.csmall.product.pojo.entity.Brand;
 import com.tqq.csmall.product.pojo.param.BrandAddNewParam;
 import com.tqq.csmall.product.pojo.param.BrandUpdateInfoParam;
@@ -50,6 +49,17 @@ public class BrandServiceImpl implements IBrandService {
     @Override
     public void delete(Long id) {
         log.debug("开始处理【根据id删除品牌信息业务】，参数：{}",id);
+        QueryWrapper<Brand> queryWrapper = new QueryWrapper<>();
+        /*这里的queryWrapper.eq就相当于查询条件where id = #{id}*/
+        /*select count(*) from pms_brand where id = #{id}*/
+        queryWrapper.eq("id",id);
+        int countById = brandMapper.selectCount(queryWrapper);
+        if (countById==0){
+            String message = "删除品牌失败,品牌信息不存在！";
+            log.warn(message);
+            throw new ServiceException(message);
+        }
+
         int rows = brandMapper.deleteById(id);
         if (rows != 1){
             String message = "删除品牌信息失败，服务器忙，请稍后再试";
@@ -62,10 +72,20 @@ public class BrandServiceImpl implements IBrandService {
     public void updateBrandById(Long id, BrandUpdateInfoParam brandUpdateInfoParam) {
         log.debug("开始处理【修改品牌信息模板】的业务，ID：{}，新数据：{}", id, brandUpdateInfoParam);
 
+        QueryWrapper<Brand> queryWrapper = new QueryWrapper<>();
+        /*这里的queryWrapper.eq就相当于查询条件where id = #{id}*/
+        /*select count(*) from pms_brand where id = #{id}*/
+        queryWrapper.eq("id",id);
+        int countById = brandMapper.selectCount(queryWrapper);
+        if (countById==0){
+            String message = "修改品牌失败,品牌信息不存在！";
+            log.warn(message);
+            throw new ServiceException(message);
+        }
+
         Brand brand = new Brand();
         BeanUtils.copyProperties(brandUpdateInfoParam,brand);
         brand.setId(id);
-        brand.setGmtModified(LocalDateTime.now());
         int rows = brandMapper.updateById(brand);
         if (rows != 1){
             String message = "修改品牌信息失败，服务器忙，请稍后再试";
